Basic data table with actions

+++ src/routes/admin/+page.server.ts
@@ -4,7 +4,7 @@
 
 import { setError, superValidate } from 'sveltekit-superforms/server';
 import { fail, redirect, error } from '@sveltejs/kit';
-import { newLinkSchema } from '$lib/formSchemas';
+import { deleteLinkSchema, newLinkSchema } from '$lib/formSchemas';
 import type { PageServerLoad } from './$types.js';
 import type { Shorturl } from '$lib/databaseItem.types.js';
 
@@ -18,7 +18,7 @@ export const actions = {
 
 		throw redirect(303, '/admin/login');
 	},
-	newlink: async ({ request, locals: { supabase } }) => {
+	create: async ({ request, locals: { supabase } }) => {
 		const form = await superValidate(request, newLinkSchema);
 		if (!form.valid) {
 			return fail(400, { form });
@@ -40,19 +40,22 @@ export const actions = {
 			return fail(500, { form });
 		}
 
-		return { open: false, success: true, action: 'create', form };
+		return { createOpen: false, success: true, action: 'create', form };
 	},
-	deletelink: async ({ request, locals: { supabase } }) => {
-		const formData = await request.formData();
-		const id = formData.get('id');
+	delete: async ({ request, locals: { supabase } }) => {
+		const form = await superValidate(request, deleteLinkSchema);
+		if (!form.valid) {
+			return fail(400, { form });
+		}
+		const { id } = form.data;
 
 		const response = await supabase.from('links').delete().eq('id', id);
 
 		if (response.error) {
-			return fail(500, { success: false, action: 'delete' });
+			return fail(500, { deleteOpen: true, success: false, action: 'delete', form });
 		}
 
-		return { success: true, action: 'delete' };
+		return { deleteOpen: false, success: true, action: 'delete', form };
 	}
 };
 
@@ -73,7 +76,8 @@ export const load: PageServerLoad = async ({ locals }) => {
 
 	const shorturls: Shorturl[] = supabaseData;
 
-	const form = await superValidate(newLinkSchema);
+	const createLinkForm = await superValidate(newLinkSchema);
+	const deleteLinkForm = await superValidate(deleteLinkSchema);
 
-	return { form, shorturls };
+	return { createLinkForm, deleteLinkForm, shorturls };
 };
